Principal Component Analysis
Principal Component Analysis on Iris data set
https://archive.ics.uci.edu/ml/datasets/iris

The dataset contains the information
gathered on three types of iris plant: Iris Setosa, Iris Versicolour, and Iris Virginica. To see if different types of Iris plants are distinguishable from one another, we can just visualize our dataset. The Iris dataset is a four dimensional dataset; therefore, it cannot be visualized in 2D or 3D. We have applied PCA to the dataset and reduce the dimensionality to two. And also colored the data points on the basis of their plant type.